Deleted unnecessary files Refactoring game file into other files. Amended .gitignore
